+++ Extension/src/background/services/rules-limits/interface.ts
@@ -31,6 +31,16 @@ export interface IRulesLimits {
     dynamicRulesMaximumCount: number;
 
     /**
+     * How many dynamic **unsafe** rules are enabled in the browser.
+     */
+    dynamicRulesUnsafeEnabledCount: number;
+
+    /**
+     * Maximum count of dynamic **unsafe** rules which can be enabled in the browser.
+     */
+    dynamicRulesUnsafeMaximumCount: number;
+
+    /**
      * How many user regexp rules are enabled in the browser.
      */
     dynamicRulesRegexpsEnabledCount: number;
@@ -114,6 +124,11 @@ type InvalidResultData = {
      * Rules count data.
      */
     rulesCount?: CountData;
+
+    /**
+     * Unsafe rules count data.
+     */
+    rulesUnsafeCount?: CountData;
 
     /**
      * Regex rules count data.